+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2TagAutoCompleteQueryDAO.java
@@ -27,10 +27,14 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
+import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
 import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.TagAutocompleteData;
 import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.TagType;
+import org.apache.skywalking.oap.server.core.query.input.Duration;
 import org.apache.skywalking.oap.server.core.storage.query.ITagAutoCompleteQueryDAO;
 import org.apache.skywalking.oap.server.library.client.jdbc.hikaricp.JDBCHikariCPClient;
+
+import static java.util.Objects.nonNull;
 
 @RequiredArgsConstructor
 public class H2TagAutoCompleteQueryDAO implements ITagAutoCompleteQueryDAO {
@@ -39,15 +43,14 @@ public class H2TagAutoCompleteQueryDAO implements ITagAutoCompleteQueryDAO {
     @Override
     public Set<String> queryTagAutocompleteKeys(final TagType tagType,
                                                 final int limit,
-                                                final long startSecondTB,
-                                                final long endSecondTB) throws IOException {
+                                                final Duration duration) throws IOException {
         StringBuilder sql = new StringBuilder();
         List<Object> condition = new ArrayList<>(2);
 
         sql.append("select distinct ").append(TagAutocompleteData.TAG_KEY).append(" from ")
            .append(TagAutocompleteData.INDEX_NAME).append(" where ");
         sql.append(" 1=1 ");
-        appendTagAutocompleteCondition(tagType, startSecondTB, endSecondTB, sql, condition);
+        appendTagAutocompleteCondition(tagType, duration, sql, condition);
         sql.append(" limit ").append(limit);
         try (Connection connection = h2Client.getConnection()) {
             ResultSet resultSet = h2Client.executeQuery(connection, sql.toString(), condition.toArray(new Object[0]));
@@ -65,14 +68,13 @@ public class H2TagAutoCompleteQueryDAO implements ITagAutoCompleteQueryDAO {
     public Set<String> queryTagAutocompleteValues(final TagType tagType,
                                                   final String tagKey,
                                                   final int limit,
-                                                  final long startSecondTB,
-                                                  final long endSecondTB) throws IOException {
+                                                  final Duration duration) throws IOException {
         StringBuilder sql = new StringBuilder();
         List<Object> condition = new ArrayList<>(3);
         sql.append("select * from ").append(TagAutocompleteData.INDEX_NAME).append(" where ");
         sql.append(TagAutocompleteData.TAG_KEY).append(" = ?");
         condition.add(tagKey);
-        appendTagAutocompleteCondition(tagType, startSecondTB, endSecondTB, sql, condition);
+        appendTagAutocompleteCondition(tagType, duration, sql, condition);
         sql.append(" limit ").append(limit);
 
         try (Connection connection = h2Client.getConnection()) {
@@ -88,16 +90,20 @@ public class H2TagAutoCompleteQueryDAO implements ITagAutoCompleteQueryDAO {
     }
 
     private void appendTagAutocompleteCondition(final TagType tagType,
-                                                final long startSecondTB,
-                                                final long endSecondTB,
+                                                final Duration duration,
                                                 final StringBuilder sql,
                                                 final List<Object> condition) {
+        long startTB = 0;
+        long endTB = 0;
+        if (nonNull(duration)) {
+            startTB = TimeBucket.getMinuteTimeBucket(duration.getStartTimestamp());
+            endTB = TimeBucket.getMinuteTimeBucket(duration.getEndTimestamp());
+        }
+
         sql.append(" and ");
         sql.append(TagAutocompleteData.TAG_TYPE).append(" = ?");
         condition.add(tagType.name());
 
-        long startTB = startSecondTB / 1000000 * 10000;
-        long endTB = endSecondTB / 1000000 * 10000 + 9999;
         if (startTB > 0) {
             sql.append(" and ");
             sql.append(TagAutocompleteData.TIME_BUCKET).append(">=?");